Debug Dijsktra

+++ src/dijkstra/PiInterface.java
@@ -1,7 +1,7 @@
 package dijkstra;
 
 public interface PiInterface {
-	public Float getDistance(VertexInterface u);
-	public void setDistance(VertexInterface u, Float i);
+	public Integer getDistance(VertexInterface u);
+	public void setDistance(VertexInterface u, Integer i);
 	public VertexInterface getClosest(VertexInterface u);
 }
